+++ cm-mlops/automation/experiment/tests/test__yaml.sh
@@ -1,0 +1,1 @@
+cm run experiment --tags=test @test_input.yaml -- {{CD}}/test_run.sh