+++ cm-mlops/script/gui/app.py
@@ -57,9 +57,9 @@ def main():
     # Set title
     if title=='':
         if script_alias!='':
-            title = 'GUI for the CK2 (CM) script "{}"'.format(script_alias)
+            title = 'GUI for the CM script "{}"'.format(script_alias)
         else:
-            title = 'GUI for CK2 (CM)'
+            title = 'GUI for CM'
 
     st.title(title)
 
@@ -288,7 +288,7 @@ def main():
         os.system(cmd2)
 
     st.markdown("""---""")
-    st.markdown("*Join the [open MLCommons taskforce](https://github.com/mlcommons/ck/blob/master/docs/mlperf-education-workgroup.md) and the public [Discord channel](https://discord.gg/JjWNWXKxwT) to participate in collaborative developments, benchmarking and design space exploration of ML Systems!"
+    st.markdown("*Join the [open MLCommons taskforce](https://github.com/mlcommons/ck/blob/master/docs/taskforce.md) and the public [Discord channel](https://discord.gg/JjWNWXKxwT) to participate in collaborative developments, benchmarking and design space exploration of ML Systems!"
                 " Please report issues or suggest features [here](https://github.com/mlcommons/ck/issues).*")
 
 if __name__ == "__main__":