+++ cm-mlops/script/print-hello-world/run.sh
@@ -3,6 +3,7 @@
 echo ""
 echo "CM_ENV_TEST1 = ${CM_ENV_TEST1}"
 echo "CM_ENV_TEST2 = ${CM_ENV_TEST2}"
+echo "CM_ENV_TEST3 = ${CM_ENV_TEST3}"
 
 echo ""
 echo "HELLO WORLD!"